Change relation between Value<T>, Value<T&> and Value<T*>
These generic Value specializations can be used to push/read user type values.
When should you use which?

Use Value<T> to copy a value onto or from the stack.
Use Value<T&> to construct or reference a value on the stack.
Use Value<T*> to copy a value (using a reference to it) onto or reference a value on the stack.